Refactor privileged trait into own code module
