+++ app/uk/gov/hmrc/individualsdetailsapi/controllers/CommonController.scala
@@ -15,18 +15,9 @@
  */
 
 package uk.gov.hmrc.individualsdetailsapi.controllers
-
 import javax.inject.Inject
 import play.api.mvc.{ControllerComponents, Result}
-import uk.gov.hmrc.auth.core.authorise.Predicate
-import uk.gov.hmrc.auth.core.retrieve.v2.Retrievals
-
-import uk.gov.hmrc.auth.core.{
-  AuthorisationException,
-  AuthorisedFunctions,
-  Enrolment
-}
-
+import uk.gov.hmrc.auth.core.AuthorisationException
 import uk.gov.hmrc.http.{HeaderCarrier, TooManyRequestException}
 
 import uk.gov.hmrc.individualsdetailsapi.domains.{
@@ -39,9 +30,6 @@ import uk.gov.hmrc.individualsdetailsapi.domains.{
 
 import uk.gov.hmrc.play.bootstrap.controller.BackendController
 
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-
 abstract class CommonController @Inject()(
     cc: ControllerComponents
 ) extends BackendController(cc) {
@@ -53,40 +41,6 @@ abstract class CommonController @Inject()(
     case _: TooManyRequestException => ErrorTooManyRequests.toHttpResponse
     case e: IllegalArgumentException =>
       ErrorInvalidRequest(e.getMessage).toHttpResponse
-  }
-}
-
-trait PrivilegedAuthentication extends AuthorisedFunctions {
-
-  val environment: String
-
-  def authPredicate(scopes: List[String]): Predicate = {
-
-    scopes.map(Enrolment(_): Predicate).reduce(_ and _)
-
-  }
-
-  def requiresPrivilegedAuthentication(scope: String)(body: => Future[Result])(
-      implicit hc: HeaderCarrier): Future[Result] = {
-
-    if (environment == Environment.SANDBOX) body
-    else authorised(Enrolment(scope))(body)
-
-  }
-
-  def requiresPrivilegedAuthenticationWithScopes(endpointScopes: List[String])(
-      implicit hc: HeaderCarrier): Future[List[String]] = {
-
-    if (environment == Environment.SANDBOX)
-      Future.successful(endpointScopes)
-    else {
-      authorised(authPredicate(endpointScopes))
-        .retrieve(Retrievals.allEnrolments) {
-          case scopes =>
-            Future.successful(scopes.enrolments.map(e => e.key).toList)
-        }
-    }
-
   }
 }
